Map enemy spawner. CreatureNode share navigation

+++ projects/Layer 2 - Application/Logia/vampirekiller.logia/commands/CommandProcessTick.cs
@@ -7,7 +7,7 @@ using Util.communication.commands;
 
 namespace vampirekiller.espeon.commands;
 
-public record struct CommandProcessTick(float delta) : ICommand
+public record struct CommandProcessTick(double delta) : ICommand
 {
     public byte[] serialize()
     {
